subject velocity

+++ src/simulation/objects/Attractor.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 
 public class Attractor extends ForceCaster{
 
-    public Attractor(int xPos, int yPos, float mass){
+    public Attractor(float xPos, float yPos, float mass){
         super(xPos, yPos, mass);
 
         try {
